+++ Arduino/DRIVERS/BEZIER_H.h
@@ -21,7 +21,7 @@ public:
     BEZIER(/* args */);
     ~BEZIER();
 
-    bool robotica_test(int = 1);
+    bool robotica_test(double = 1);
     bool circle_segment_offr(double, double, double, double, double);
     bool circle_segment_onr(double, double, double);
     //bool random_curve();
